grades fix

+++ projects/average_grade.py
@@ -5,6 +5,6 @@ average = 0
 for i in range(1, int(periods) + 1):
     grades.append(input(f"percentage grade in class {i}? > "))
 for i in grades:
-    average = average + int(i)
+    average = average + float(i)
 average = round(average / len(grades), 2)
 print(f"your average grade is {average}.")
